# Make parsing of `.amr` files faster