This PR speeds up the parsing of the `amr` files by using `pandas` directly because I/O in astropy can be slow for particular kinds of files. This also speeds up the indexing of `Strand` by passing an index explicitly to `Profile` in `Strand.__getitem__` when possible.

Before this PR:

```python
from pydrad.parse.util import read_amr_file
%%timeit
read_amr_file(test_filename)
12.9 ms ± 70.6 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

```python
from pydrad.parse import Strand
s = Strand(test_dir, read_phy=False, read_ine=False, read_trm=False, read_hstate=False, read_scl=False)
%%timeit
s[10]
14.6 ms ± 53 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

After this PR:

```python
%%timeit
read_amr_file(test_filename)
1.31 ms ± 8.04 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

```python
%%timeit
s[10]
1.52 ms ± 134 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```